+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- You can pass variable arguments to System.run()
+
+### Deprecated
+
+- Assign variable arguments to System()
+
 ## [2017-06-08] - 1.2.0
 
 ### Added